+++ CHANGELOG.md
@@ -7,9 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Added logging for issues with creating resource folders
+- Included Mockito for testing
 
 ### Changed
 - Updated resource loader to use SLF4j string formatting
+- Changed plugin commands to use lists instead of arrays
 
 ## [0.5.0] - 2023-05-31
 ### Added